// src/app/models/site-info.model.ts
export interface SiteInfo {
    id?: number;
    site_id: number;
    site_name?: string;
    block_no?: string;
    street_no?: string;
    house_no?: string;
    date_started?: string;
    date_ended?: string;
    items_data?: any;
    attendance_data?: any;
    receipts?: string[];
    created_at?: string;
    updated_at?: string;
}